feat(packages/explorer/src/views/Mining/): Changed view to show progress bar with metamask installed
Changed view to show relevant information without metamask installed

fix #180
