+++ packages/explorer/src/views/Mining/MiningArea.js
@@ -1,0 +1,92 @@
+import React from 'react'
+import TableData from './TableData'
+import { Button } from '../../components'
+
+const MiningArea = ({
+  amtLpt,
+  balance,
+  estimCost,
+  gas,
+  handleSubmit,
+  loading,
+  lowBal,
+  progress,
+}) => {
+  return (
+    <React.Fragment>
+      {!loading && !progress && <h1>Your mining parameters</h1>}
+      {!progress && (
+        <React.Fragment>
+          <table>
+            <tbody>
+              <tr>
+                <TableData
+                  tableData="Account balance:"
+                  help="The balance on your web3 enabled browser or wallet plugin."
+                />
+                <TableData beforeData={balance.toString()} tableData="Ether" />
+              </tr>
+              {lowBal && (
+                <tr>
+                  <TableData
+                    tableData="You do not have sufficient funds in your web-3 wallet to mine LPT tokens."
+                    colSpan="2"
+                  />
+                </tr>
+              )}
+              <tr>
+                <TableData
+                  tableData="Gas price:"
+                  help="The current market price of 1 gas according to EthGasStation."
+                />
+                <TableData tableData="Gwei" beforeData={gas} />
+              </tr>
+            </tbody>
+          </table>
+          <hr />
+          <h1>Estimated mining results</h1>
+          <table>
+            <tbody>
+              <tr>
+                <TableData
+                  tableData="Estimated cost:"
+                  help={`Total total cost of one round of mining and generating
+                          tokens for the 20 eligible ethereum addresses.`}
+                />
+                <TableData beforeData={estimCost} tableData="Ether" />
+              </tr>
+              <tr>
+                <TableData
+                  tableData="Estimated time:"
+                  help={`The time it takes to complete one round of mining.`}
+                />
+                <TableData beforeData="1 - 5" tableData="Min" />
+              </tr>
+              <tr>
+                <TableData
+                  tableData="Estimated number of LPT tokens you will earn:"
+                  help={`The portion of the LPT tokens that will be issued to
+                          you in one round of mining.`}
+                />
+                <TableData beforeData={amtLpt} tableData="LPT" />
+              </tr>
+            </tbody>
+          </table>
+          <div className="center-div">
+            <Button
+              onClick={handleSubmit}
+              style={{
+                backgroundColor: lowBal ? '#ccc' : '#000000',
+              }}
+              className="primary-btn"
+            >
+              Earn LPT
+            </Button>
+          </div>
+        </React.Fragment>
+      )}
+    </React.Fragment>
+  )
+}
+
+export default MiningArea
