+++ packages/explorer/src/views/Mining/NoTokens.js
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const NoTokensPage = () => {
+  return (
+    <React.Fragment>
+      <h1>All tokens have been claimed.</h1>
+      <p>There are no remaining LPT tokens to mine.</p>
+      <p>
+        If you are interested in obtaining LPT tokens, visit one of the
+        decentralized exchanges that currently provide liquidy for LPT tokens.
+      </p>
+    </React.Fragment>
+  )
+}
+
+export default NoTokensPage
